+++ cmsafvis/R/getColors.R
@@ -13,6 +13,10 @@ getColors <- function(PAL,
   }
   curPAL <- as.list(palettes[idx, ])
   if (length(idx) == 0) {
+    idx <- which(rownames(palettes) == "larry")
+    name   <- "larry"
+    curPAL <- as.list(palettes[idx, ])
+    
     idx <- which(rownames(palettes) == "sunny")
     name   <- "sunny"
     curPAL <- as.list(palettes[idx, ])
@@ -21,6 +25,19 @@ getColors <- function(PAL,
   if (curPAL$type == "base") {
     pal <- eval(parse(text = tolower(name)))
   } else if (curPAL$type == "more") {
+    larry <- grDevices::colorRampPalette(c("#023858",
+                                           "#0570B0",
+                                           "#6EAAC8",
+                                           "#53BD9F",
+                                           "#99F0B2",
+                                           "#CDFFCD",
+                                           "#FFFFFF",
+                                           "#FFF5BA",
+                                           "#F5E09E",
+                                           "#F5CD84",
+                                           "#E1A564",
+                                           "#CD853F",
+                                           "#B66A28"))
     # somehow sunny IS used
     sunny <- grDevices::colorRampPalette(c("black",
                                            "#3a0303",