+++ cmsafvis/NEWS.md
@@ -5,6 +5,7 @@
 - plotKML package is back on CRAN und export of KML is possible again
 - fieldmean plot includes now all available years and not just years included
   in the climatology period
+- new color scale 'larry' added
 
 ## 1.1.8